Fix build
Currently, due to some issues such as
https://youtrack.jetbrains.com/issue/KT-46153,

It's only possible to publish artifacts for other platforms (such as
Desktop or IOS) if all the dependencies have the specific target as well

Currently, Kotest doesn't support IOS (it's a work in progress), and
thus this extension can't support it either.

This commit removes these targets until we're ready to support them. The
only targets left are JVM and JS.